refactor(gspenst): lets user run tinacms cli

+++ packages/gspenst/src/tinaServer.ts
@@ -7,7 +7,8 @@ import { createLogger } from './logger'
 const log = createLogger('tinaServer')
 
 export async function startTinaServer(
-  this: { projectPath: string } & EventEmitter
+  this: { projectPath: string } & EventEmitter,
+  config: { onlyCheck: boolean }
 ) {
   if (
     !(await fse.pathExists(
@@ -17,30 +18,34 @@ export async function startTinaServer(
     throw new Error('Theme is missing schema definition')
   }
 
-  log('Starting tina server', this.projectPath)
+  if (!config.onlyCheck) {
+    log('Starting tina server', this.projectPath)
 
-  // TODO remove `--noWatch` and add `--experimentalData`
-  const ps = spawn('tinacms', ['server:start', '--noWatch'], {
-    cwd: this.projectPath,
-  })
+    // TODO remove `--noWatch` and add `--experimentalData`
+    const ps = spawn('tinacms', ['server:start', '--noWatch'], {
+      cwd: this.projectPath,
+    })
 
-  this.on('cleanup', () => {
-    ps.kill()
-  })
+    this.on('cleanup', () => {
+      ps.kill()
+    })
 
-  if (ps.stdout) {
-    let flag = true
-    while (flag) {
-      const [data] = (await once(ps.stdout, 'data')) as Buffer[] // eslint-disable-line no-await-in-loop
-      const msg = data?.toString().trim()
-      if (msg) {
-        log(msg)
-        if (msg.includes('4001')) {
-          flag = false
+    // TODO output errors
+    if (ps.stdout) {
+      let flag = true
+      while (flag) {
+        const [data] = (await once(ps.stdout, 'data')) as Buffer[] // eslint-disable-line no-await-in-loop
+        const msg = data?.toString().trim()
+        if (msg) {
+          log(msg)
+          // eslint-disable-next-line max-depth
+          if (msg.includes('4001')) {
+            flag = false
+          }
         }
       }
     }
-  }
 
-  return ps
+    return ps
+  }
 }
